# API/Sesame/src/main/java/com/PDL/Sesame/model/NatureQuestion.java
package com.PDL.Sesame.model;

import com.fasterxml.jackson.annotation.JsonIgnore;
import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import com.fasterxml.jackson.annotation.JsonManagedReference;
import jakarta.persistence.Entity;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import jakarta.persistence.*;

import java.util.ArrayList;
import java.util.List;

@Entity
@Data
@NoArgsConstructor
@AllArgsConstructor
public class NatureQuestion {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id_nature_question;

    @Column(nullable = false)
    private String nom_nature_question;

    @JsonIgnore
    //@JsonIgnoreProperties("nature")
    @OneToMany(mappedBy = "nature")
    private List<Question> questions ;

}